+++ core/f_time.cpp
@@ -6,10 +6,6 @@
 
 using namespace Fresa;
 
-namespace {
-    TimerID last_timer_id = 0;
-}
-
 Clock::time_point Fresa::time() {
     //---The actual time ^·^---
     return Clock::now();
@@ -18,11 +14,9 @@ Clock::time_point Fresa::time() {
 TimerID Fresa::setTimer(ui32 ms) {
     //---Set timer---
     //      Creates a new timer and appends it to the timer list, saves the duration and start time
-    TimerID timer_id = last_timer_id + 1;
-    
-    while (Time::timers.find(timer_id) != Time::timers.end())
-        timer_id++;
-    last_timer_id = timer_id;
+    static TimerID timer_id = 0;
+    do timer_id++;
+        while (Time::timers.find(timer_id) != Time::timers.end());
     
     Time::timers[timer_id] = Timer(std::chrono::milliseconds(ms), time());